refactor: modularize stock management logic by migrating helper functions from api.py to kta_stock managers and cleaning up unused code.
